pyproject.toml
[project]
name = "spa-sql-supporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "flask>=3.1.3",
    "oracledb>=3.4.2",
    "paramiko>=4.0.0",
    "python-dotenv>=0.21.0",
    "pyyaml>=6.0.3",
    "requests>=2.0",
    "sqlfluff>=4.0.4",
]
